MAGETWO-57410: Adding too many configurable product option causes browser to hang.
 - a better way to serialize configurable-matrix and associate_product_ids when editing a configurable product with many variations
